+++ actions/excel/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [4.0.2] - 2025-03-06
+
+### Changed
+
+- Dependency versions updated
+- Use `Response[T]` as return type to actions
+
 ## [4.0.1] - 2025-02-17
 
 ### Fixed