+++ actions/microsoft-sharepoint/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.0.2] - 2025-03-06
+
+### Changed
+
+- Dependency versions updated
+
 ## [2.0.1] - 2025-01-09
 
 ### Changed